requirements.txt
Django==6.0.2
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-jazzmin==3.0.1
djangorestframework==3.15.2
drf-spectacular==0.28.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.3
whitenoise==6.9.0
python-telegram-bot==21.10
aiohttp==3.11.12
pillow==11.1.0
asgiref>=3.9.1
dj-database-url==2.3.0
deep-translator==1.11.4